requirements.txt
flask
requests
flask-cors
python-dotenv
google.generativeai
pillow
PyGithub
firebase-admin
